safari-600.7-branch/LayoutTests/fast/dom/shadow/shadow-dom-event-dispatching-manually-fired-with-same-related-target-expected.txt

Tests to ensure that event dispatching behaves as the Shadow DOM spec describes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




Composed Shadow Tree will be:
DIV	 id=sandbox
	DIV	 id=top
		DIV	 id=host
			DIV	 id=div1


  mouseout

  mouseover
     @div1 (target: div1) (related: div1)
     @shadow-root (target: div1) (related: div1)
PASS successfullyParsed is true

TEST COMPLETE
